fix: render Latin Extended-A glyphs in map names, embed camera below HUD
Map name rendering stored char codes in Uint8 buffers and an R8UI
texture, folding codepoints above 255 onto unrelated glyphs (U+0142
lstroke rendered as B, U+0119 eogonek vanished) even though the MSDF
atlas and glyph tables (CHAR_RANGE 384) already cover Latin Extended-A.
Widened the string pipeline to Uint16/R16UI and codepoints outside the
tables now fall back to ? instead of aliasing. In hexah-embedded mode
the initial camera fits the map below the top HUD row (96px) so the
leaderboard and game controls no longer cover it.

+++ src/client/render/gl/passes/WorldTextPass.ts
@@ -114,7 +114,7 @@ export class WorldTextPass {
   private kernTable: Int8Array;
 
   // Reusable buffers for layoutString
-  private charCodes = new Uint8Array(MAX_CHARS);
+  private charCodes = new Uint16Array(MAX_CHARS);
   private cursors = new Float32Array(MAX_CHARS);
 
   private distanceRange: number;
